fix: ETU-71360 use consistent 'long' type for 'outage-tolerant' value

+++ oidc-rs-spring-boot-common/src/main/java/org/entur/auth/spring/common/server/IssuerProperties.java
@@ -17,5 +17,5 @@ public class IssuerProperties {
     private Integer refreshAheadTime;
     private Integer cacheRefreshTimeout;
     private Integer cacheLifespan;
-    private Integer outageTolerant;
+    private Long outageTolerant;
 }
